requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
pytest-django>=4.5.0
pydantic>=2.0.0
pydantic[email]>=2.0.0 